+++ src/software/ai/hl/stp/play/defense/defense_play_base.cpp
@@ -16,7 +16,7 @@ void DefensePlayFSMBase::setUpCreaseDefenders(int num_crease_defenders)
     }
 
     crease_defenders =
-        std::vector<std::shared_ptr<CreaseDefenderTactic>>(num_crease_defenders);
+        std::vector<std::shared_ptr<CreaseDefenderTactic>>(static_cast<std::size_t>(num_crease_defenders));
     std::generate(crease_defenders.begin(), crease_defenders.end(),
                   [this]() { return std::make_shared<CreaseDefenderTactic>(ai_config); });
 }
@@ -28,7 +28,7 @@ void DefensePlayFSMBase::setUpPassDefenders(int num_pass_defenders)
         return;
     }
 
-    pass_defenders = std::vector<std::shared_ptr<PassDefenderTactic>>(num_pass_defenders);
+    pass_defenders = std::vector<std::shared_ptr<PassDefenderTactic>>(static_cast<std::size_t>(num_pass_defenders));
     std::generate(pass_defenders.begin(), pass_defenders.end(),
                   [this]() { return std::make_shared<PassDefenderTactic>(ai_config); });
 }
